Convert wire_assigns from dict to Bimap.
The previous code stored wire assignments as a dict of sink wires to a
list of source wires.  However there are many instances where the
reverse lookup was required, and this was resulting in a significant
number of linear scans that were very slow.

WireAssignsBimap replaces the old dict -> list with a dict -> set in
each direction.

Signed-off-by: Keith Rothman <537074+litghost@users.noreply.github.com>